refactor(wework): move native context menu wiring into host actions module
Keeps main.ts focused and lets real-Electron verification import the
exact production wiring (createNativeContextMenuActions +
installContextMenu) without booting the whole app.
